SENTRY-2124: LeaderStatusMonitor.toString() throws IllegalFormatConversionException with AtomicLong (Xinran Tinney, reviewed by Sergio Pena)
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/30845
Tested-by: Jenkins User

+++ sentry-provider/sentry-provider-db/src/main/java/org/apache/sentry/service/thrift/LeaderStatusMonitor.java
@@ -283,7 +283,7 @@ final class LeaderStatusMonitor
   public String toString() {
     return isSingleNodeMode?"Leader election disabled":
         String.format("{isSingleNodeMode=%b, incarnationId=%s, isLeader=%b, leaderCount=%d}",
-        isSingleNodeMode, incarnationId, isLeader, leaderCount);
+        isSingleNodeMode, incarnationId, isLeader, leaderCount.longValue());
   }
 
 }
